moved mat read tools to sbximagingextractor where they are used.
